% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_clusters.R
\name{clickedClusterSum}
\alias{clickedClusterSum}
\title{Nearest clicked cluster}
\usage{
clickedClusterSum(corpus = NULL, e = NULL)
}
\arguments{
\item{corpus}{a document corpus}

\item{e}{co-ordinates from mouse click}
}
\description{
Nearest clicked cluster
}
