% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_clusters.R
\name{clusterSummaryText}
\alias{clusterSummaryText}
\title{Print a summary statement for topic clusters}
\usage{
clusterSummaryText(corpus = NULL, clustSelected = NULL)
}
\arguments{
\item{corpus}{a document corpus}

\item{clustSelected}{name of a cluster}
}
\description{
Print a summary statement for topic clusters
}
