% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_singleFile.R
\name{genID}
\alias{genID}
\title{genID
generic id generating function}
\usage{
genID(n = NULL, idLen = 10)
}
\arguments{
\item{n}{}

\item{idLen}{}
}
\description{
genID
generic id generating function
}
