% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_clusters.R
\name{getTopPapers}
\alias{getTopPapers}
\title{getTopPapers}
\usage{
getTopPapers(corpus = NULL, selectedCluster = NULL)
}
\arguments{
\item{corpus}{a document corpus}

\item{selectedCluster}{a name of cluster}
}
\description{
Top Papers for a cluster
}
