% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_clusters.R
\name{getTopTerms}
\alias{getTopTerms}
\title{Naming clusters}
\usage{
getTopTerms(clustPMID = NULL, topNVal = 1, clustValue = NA,
  tidyCorpus = NULL)
}
\arguments{
\item{clustPMID}{}

\item{topNVal}{}

\item{clustValue}{}

\item{tidyCorpus}{}
}
\value{
topWord : a string of top two terms for a given cluster
}
\description{
Naming clusters
}
\examples{
See online useage demonstration:https://github.com/amcrisan/Adjutant#demo
}
