% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_singleFile.R
\name{process_inputFile}
\alias{process_inputFile}
\title{process_inputFile
a generic method for reading in differnt text flat files. Helper method for process_singleFile}
\usage{
process_inputFile(fIN = NULL)
}
\arguments{
\item{fIN}{}
}
\value{
dat
}
\description{
process_inputFile
a generic method for reading in differnt text flat files. Helper method for process_singleFile
}
