% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_semanticScholar.R
\name{processScholarSearch}
\alias{processScholarSearch}
\title{Process Scholar Search
Method that will also explore Semantic Scholar Articles}
\usage{
processScholarSearch(
  query = NULL,
  query_max = 40000,
  totalPages = 1000,
  yearMin = NULL,
  yearMax = NULL,
  publicationTypes = list(),
  requirePDF = TRUE,
  sortResults = "relevance"
)
}
\arguments{
\item{query}{}
}
\value{

}
\description{
Process Scholar Search
Method that will also explore Semantic Scholar Articles
}
