% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pubmed.R
\name{processSearch}
\alias{processSearch}
\title{Processing PubMed Search}
\usage{
processSearch(query = NULL, demoversion = FALSE, ...)
}
\arguments{
\item{query}{query for PubMed search}

\item{demoversion}{hard limit on PubMED retrieved records to 1000. [Default: FALSE]}

\item{...}{arguements passed to RISmed::EUtilsSummary method}
}
\value{
corpus
}
\description{
Wrapper interface for RISmed and adding additional article metadata.
}
\examples{
See online useage demonstration:https://github.com/amcrisan/Adjutant#demo
}
