% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_singleFile.R
\name{processSingleFile}
\alias{processSingleFile}
\title{process_singleFile
A function that allows a user to run adjutant on a single file that has nothing to do with a lit review.}
\usage{
processSingleFile(dat = NULL, colCollapse = NULL)
}
\arguments{
\item{dat}{}

\item{colCollapse}{}
}
\value{
dat
}
\description{
process_singleFile
A function that allows a user to run adjutant on a single file that has nothing to do with a lit review.
}
