% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_corpus.R
\name{runTSNE}
\alias{runTSNE}
\title{Wrapped for rTSNE}
\usage{
runTSNE(tidyCorpus_df = NULL, ...)
}
\arguments{
\item{tidyCorpus_df}{A tidy corpus. Run the "tidyCorpus" method to generate the expected input}

\item{...}{parameters that can be passed to rTSNE}
}
\value{
a data frame with with x,y t-SNE co-ordinates for
}
\description{
Wrapped for rTSNE
}
\details{
A wrapper for the t-SNE method that creates a document term matrix from a tidy corpus, run t-SNE, and provides a cleaned up data frame output for further analysis.
}
