% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_corpus.R
\name{tidyCorpus}
\alias{tidyCorpus}
\title{Tidying Up the PubMed Corpus}
\usage{
tidyCorpus(corpus = NULL)
}
\arguments{
\item{corpus}{a document corpus}
}
\value{
Tidy Text Corpus
}
\description{
Creates a tidytext corpus from the PubMed articles titles and abstracts. Adjutant generally follows the approach presented in  'Text Mining with R' (https://www.tidytextmining.com/tidytext.html).
}
\examples{
\dontrun{See online useage demonstration:https://github.com/amcrisan/Adjutant#demo}
}
