% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_clusters.R
\name{topClustTerms}
\alias{topClustTerms}
\title{Get Top Terms in Each Cluster}
\usage{
topClustTerms(corpus = NULL, corpusTidy = NULL,
  selectedCluster = NULL)
}
\arguments{
\item{corpus}{a document corpus}

\item{corpusTidy}{a tidy text version of the corpus}

\item{selectedCluster}{a name of a cluster}
}
\description{
Get Top Terms in Each Cluster
}
