% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_clusters.R
\name{getTopTerms}
\alias{getTopTerms}
\title{Naming clusters}
\usage{
getTopTerms(clustPMID = NULL, topNVal = 1, clustValue = NA,
  tidyCorpus = NULL)
}
\arguments{
\item{clustPMID}{PMID values for documents within the cluster}

\item{topNVal}{# of top terms used to name cluster}

\item{clustValue}{cluster identified, mainly used to check it unclustered}

\item{tidyCorpus}{tidytext document corpus}
}
\value{
topWord : a string of top two terms for a given cluster
}
\description{
Naming clusters
}
\examples{
\dontrun{See online useage demonstration:https://github.com/amcrisan/Adjutant#demo}
}
