% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_clusters.R
\name{topClustTerms}
\alias{topClustTerms}
\title{Get Top Terms in Each Cluster}
\usage{
topClustTerms(corpus = NULL, corpusTidy = NULL, selectedCluster = NULL)
}
\description{
Get Top Terms in Each Cluster
}
